Change types from float to double

+++ modules/imgproc/test/test_resize_nn.cpp
@@ -10,7 +10,7 @@ namespace opencv_test { namespace {
         {
             int cols = 12, rows = 12;
             const int N = 10;
-            float scale_factor = 0.9;
+            double scale_factor = 0.9;
             int i1 = 5, j1 = 5; // position of pixel
             for (int i = 2; i < N; ++i)
             {
@@ -21,7 +21,7 @@ namespace opencv_test { namespace {
 
                 cv::resize(test, dst, Size(0, 0), scale_factor, scale_factor, cv::INTER_NEAREST);
                 dst = dst.reshape(1, dst.cols * dst.rows * dst.channels());
-                float sum = cv::sum(dst)[0];
+                double sum = cv::sum(dst)[0];
                 EXPECT_EQ(i, sum) << "Resize NN TEST mat from " << cols << "x" << rows << "x" << i << " failed with sum " << sum << " should be " << i;
             }
         }
